Handling no database config

+++ src/server/storage/banStorage.ts
@@ -26,6 +26,7 @@ function dbToBan(banId: string, dbBan: db.Ban): s.Ban {
 
 export async function loadAllBans(): Promise<s.Ban[]> {
     const firestore = getFirestore();
+    if (!firestore) { return []; }
 
     const bans = new Array<s.Ban>();
     const querySnapshot = await firestore.collection(Collections.Ban).get();
@@ -39,12 +40,16 @@ export async function loadAllBans(): Promise<s.Ban[]> {
 
 export async function createOrUpdateBan(ban: s.Ban) {
     const firestore = getFirestore();
+    if (!firestore) { return; }
+
     const doc = firestore.collection(Collections.Ban).doc(ban.banId);
     await doc.set(banToDb(ban));
 }
 
 export async function deleteBan(banId: string) {
     const firestore = getFirestore();
+    if (!firestore) { return; }
+
     const doc = firestore.collection(Collections.Ban).doc(banId);
     await doc.delete();
 }
